pyproject.toml
[project]
name = "auto-leetcode"
version = "0.1.0"
description = "Automatically solve LeetCode problems using AI"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27",
    "click>=8.1",
    "python-dotenv>=1.0",
    "openai>=1.30",
    "anthropic>=0.30",
    "beautifulsoup4>=4.12",
]

[project.scripts]
auto-leetcode = "auto_leetcode.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "black>=24.0",
    "ruff>=0.4",
    "isort>=5.13",
    "mypy>=1.10",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM"]

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
strict = true
